join button kitchensink to .stories.tsx
